cutover to utils.client.py
